# docs/swagger.yaml
info:
  contact: {}
paths:
  /todos:
    get:
      consumes:
      - application/json
      description: Get all todos
      produces:
      - application/json
      responses:
        "500":
          description: Internal Server Error
          schema: {}
      summary: Get all todos
      tags:
      - Todos
    post:
      consumes:
      - application/json
      description: Create a new todo
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Create a new todo
      tags:
      - Todos
  /todos/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - text/plain
      responses:
        "204":
          description: No Content
        "500":
          description: Internal Server Error
          schema: {}
      summary: Delete a todo
      tags:
      - Todos
    get:
      consumes:
      - application/json
      description: Get a todo by ID
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "404":
          description: Not Found
          schema: {}
      summary: Get a todo by ID
      tags:
      - Todos
    put:
      consumes:
      - application/json
      description: Update a todo
      parameters:
      - description: Todo ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "400":
          description: Bad Request
          schema: {}
        "500":
          description: Internal Server Error
          schema: {}
      summary: Update a todo
      tags:
      - Todos
swagger: "2.0"